moved shard slot errors and process_shard_slot into own file
