+++ eth2/shard_state_processing/src/per_shard_slot_processing/errors.rs
@@ -1,0 +1,6 @@
+use types::*;
+
+#[derive(Debug, PartialEq)]
+pub enum Error {
+    ShardStateError,
+}
